squash printf warning

+++ src/type_contiguous_x.c
@@ -75,7 +75,7 @@ int MPIX_Type_contiguous_x(MPI_Aint offset, MPI_Count count, MPI_Datatype oldtyp
 {
     /* The count has to fit into MPI_Aint for BigMPI to work. */
     if ((MPI_Count)count>(MPI_Count)bigmpi_count_max) {
-        printf("count (%lld) exceeds bigmpi_count_max (%lld)\n", (long long)count, bigmpi_count_max);
+        printf("count (%lld) exceeds bigmpi_count_max (%lld)\n", (long long)count, (long long)bigmpi_count_max);
         fflush(stdout);
     }
